third_party/WebKit/LayoutTests/editing/execCommand/selection-after-switch-type-of-listitem.html

<script src="../../resources/dump-as-markup.js"></script>
<div id="test" contenteditable>
<ul>
 <li id="first">ABC</li>
 <li id="second">DEF</li>
</ul>
</div>
<script>
var selection = window.getSelection();
var range = document.createRange();
range.setStart(document.getElementById('second').firstChild, 0);
range.setEnd(document.getElementById('second').firstChild, 3);
selection.addRange(range);
document.execCommand('InsertOrderedList');
document.execCommand('InsertUnOrderedList');
Markup.description('This tests verifies that the selection is properly reset after executing the command.');
Markup.dump('test');
</script>
