third_party/blink/web_tests/platform/generic/editing/execCommand/selection-after-switch-type-of-listitem-expected.txt

This tests verifies that the selection is properly reset after executing the command.
| "
"
| <ul>
|   "
 "
|   <li>
|     id="first"
|     "ABC"
|   "
 "
|   <li>
|     "<#selection-anchor>DEF<#selection-focus>"
|   "
"
| "
"
